dbt-analyse/requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
google-cloud-bigquery>=3.11.0
pyarrow>=12.0.0
dbt-core>=1.7.0
dbt-bigquery>=1.7.0
